Use higher lower dummy game instead of stub
