+++ src/botdriver/src/main.rs
@@ -1,4 +1,5 @@
 mod types;
+mod higher_lower;
 
 extern crate serde;
 
@@ -14,6 +15,8 @@ use std::fs::File;
 use std::io::prelude::*;
 use std::path::PathBuf;
 
+use higher_lower::HigherLower;
+
 fn main() {
     let game_config = match parse_config() {
         Ok(config) => config,
@@ -22,7 +25,7 @@ fn main() {
             std::process::exit(1)
         }
     };
-    //run::<Stub>();
+    run::<HigherLower>();
 }
 
 fn parse_config() -> Result<GameConfig, Box<Error>> {
@@ -90,32 +93,5 @@ fn fetch_player_output(player: &Player, info: &GameInfo) -> PlayerCommand {
         ("Ilion", "slecht nieuws") => "paniek".to_owned(),
         ("Anna", "goed nieuws") => "de doodssteek".to_owned(),
         _ => "een beweging".to_owned()
-    }
-}
-
-struct Stub;
-
-impl Game for Stub {
-    fn init(names: Vec<Player>) -> Self {
-        Stub
-    }
-
-    fn start(&mut self) -> GameStatus {
-        GameStatus::Running(vec![
-            ("Ilion".to_owned(), "start".to_owned()),
-            ("Anna".to_owned(), "betere start".to_owned())
-        ])
-    }
-
-    fn step(&mut self, player_output: &PlayerOutput) -> GameStatus {
-        let mut pi = vec![];
-        for &(ref player, ref command) in player_output {
-            match (player.as_ref(), command.as_ref()) {
-                ("Ilion", "eerste zet") => pi.push(("Ilion".to_owned(), "slecht nieuws".to_owned())),
-                ("Anna", "betere eerste zet") => pi.push(("Anna".to_owned(), "goed nieuws".to_owned())),
-                _ => return GameStatus::Done(Outcome::Score(vec![("Anna".to_owned(), 45), ("Ilion".to_owned(), 0)]))
-            }
-        }
-        GameStatus::Running(pi)
     }
 }
